feat: Add gyroscope calibration functionality with a dedicated UI and firmware command handling.

+++ software/src/protocol/PacketDecoder.h
@@ -23,7 +23,7 @@ public:
   static QString typeToString(uint8_t type);
 
 private:
-  QString sysStateToName(uint8_t state);
+  QString sysStateToName(uint16_t state);
   ImuData m_lastImu;
   bool m_hasLastImu = false;
 };
